update outbound message interface
